/graphql?schema=mock uses mockSchema
