+++ package.json
@@ -16,6 +16,7 @@
     "graphiql": "^0.11.11",
     "graphql": "^14.0.2",
     "graphql-tools": "^3.1.1",
+    "lodash": "^4.17.11",
     "next": "^6.1.2",
     "react": "^16.5.0",
     "react-dom": "^16.5.0"
