+++ src/schema/mocks.js
@@ -1,0 +1,24 @@
+export default {
+  Query: {
+    books() {
+      return [
+        {
+          title: "Fake Title #1",
+          author: "Fake Author #1"
+        },
+        {
+          title: "Fake Title #2",
+          author: "Fake Author #2"
+        },
+        {
+          title: "Fake Title #3",
+          author: "Fake Author #3"
+        },
+        {
+          title: "Fake Title #4",
+          author: "Fake Author #4"
+        }
+      ]
+    }
+  }
+}
