+++ src/schema/typeDefs.js
@@ -1,0 +1,14 @@
+export default `
+  type Book {
+    title: String
+    author: String
+  }
+
+  type Query {
+    version: String!
+  }
+
+  extend type Query {
+    books: [Book]
+  }
+`
